requirements.txt
appdirs
aria2p
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
js2py
lxml
megasdkrestclient
psutil
Pyrogram
python-dotenv
python-magic
python-telegram-bot
requests
telegraph
tenacity
TgCrypto
youtube_dl
